Refactor project structure and update dependencies
- Updated `API.csproj` to change `UserSecretsId` and remove `Infrastructure` reference.
- Replaced `baseUrl` with `API_HostAddress` in `API.http` for consistency.
- Modified `ExchangeRateController.cs` to use `Domain.Interfaces`.
- Added new namespaces in `Program.cs` from the `Domain` layer.
- Updated `appsettings.json` with user credentials for `CentralBankService`.
- Removed entire project structure and dependencies from `Application.csproj`.
- Deleted `CentralBankServiceClient.cs` class and its methods.
- Moved `CentralBankServiceOptions.cs` to `Domain.Configuration`.
- Updated `Domain.csproj` with new package references for `System.ServiceModel`.
- Modified `ExchangeRateEntity.cs` to include exchange rate properties.
- Moved `IExchangeRateService.cs` to `Domain.Interfaces`.
- Updated `ExchangeRateService.cs` for namespace and dependency changes.
- Adjusted solution file `TipoCambioCR.sln` for project reference updates.
