+++ API/Controllers/ExchangeRateController.cs
@@ -1,4 +1,4 @@
-﻿using Application.Services;
+﻿using Domain.Interfaces;
 using Microsoft.AspNetCore.Mvc;
 using System.Globalization;
