+++ API/appsettings.json
@@ -6,7 +6,7 @@
     }
   },
   "AllowedHosts": "*",
-    "CentralBankService": {
+  "CentralBankService": {
     "Usuario": "Deyler Alvarez",
     "Correo": "deylerafernandez@gmail.com",
     "Token": "6NAN2AA77E"
