--- Domain/Entities/ErrorLog.cs
@@ -1,1 +1,0 @@
-﻿
