+++ Domain/Entities/ExchangeRateEntity.cs
@@ -1,4 +1,4 @@
-namespace Domain.Entities
+﻿namespace Domain.Entities
 {
     public class ExchangeRateEntity
     {
