+++ Domain/Interfaces/IExchangeRateService.cs
@@ -1,6 +1,6 @@
-using Domain.Entities;
+﻿using Domain.Entities;
 
-namespace Application.Services
+namespace Domain.Interfaces
 {
     public interface IExchangeRateService
     {
