--- Infrastructure/Repositories/ErrorLogRepository.cs
@@ -1,1 +1,0 @@
-﻿
