--- Infrastructure/Repositories/ExchangeRateRepository.cs
@@ -1,1 +1,0 @@
-﻿
