--- Infrastructure/Repositories/IExchangeRateRepository.cs
@@ -1,1 +1,0 @@
-﻿
