--- Infrastructure/Repositories/ILogRepository.cs
@@ -1,1 +1,0 @@
-﻿
